Added links to posts in blog page. Organized files
