+++ components/common/Markdown/index.js
@@ -1,0 +1,83 @@
+import { useMemo } from 'react';
+
+import ReactMarkdown from 'react-markdown';
+import rehypeRaw from 'rehype-raw';
+import remarkGfm from 'remark-gfm';
+import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter';
+import tsx from 'react-syntax-highlighter/dist/cjs/languages/prism/tsx';
+import typescript from 'react-syntax-highlighter/dist/cjs/languages/prism/typescript';
+import scss from 'react-syntax-highlighter/dist/cjs/languages/prism/scss';
+import bash from 'react-syntax-highlighter/dist/cjs/languages/prism/bash';
+import markdown from 'react-syntax-highlighter/dist/cjs/languages/prism/markdown';
+import json from 'react-syntax-highlighter/dist/cjs/languages/prism/json';
+
+import { useSite } from '@/components/common/Site';
+import oneDark from '@/data/syntax-dark';
+import oneLight from '@/data/syntax-light';
+
+SyntaxHighlighter.registerLanguage('tsx', tsx);
+SyntaxHighlighter.registerLanguage('typescript', typescript);
+SyntaxHighlighter.registerLanguage('scss', scss);
+SyntaxHighlighter.registerLanguage('bash', bash);
+SyntaxHighlighter.registerLanguage('markdown', markdown);
+SyntaxHighlighter.registerLanguage('json', json);
+
+const Markdown = ({ children, md }) => {
+  const { colorScheme } = useSite();
+  const syntaxTheme = useMemo(
+    () => (colorScheme === 'dark' ? oneDark : oneLight),
+    [colorScheme]
+  );
+
+  const MarkdownComponents = {
+    code({ node, inline, className, ...props }) {
+      const hasLang = /language-(\w+)/.exec(className || '');
+      const hasMeta = node?.data?.meta;
+
+      const applyHighlights = (applyHighlights) => {
+        if (hasMeta) {
+          const RE = /{([\d,-]+)}/;
+          const metadata = node.data.meta?.replace(/\s/g, '');
+          const strlineNumbers = RE?.test(metadata)
+            ? RE?.exec(metadata)[1]
+            : '0';
+          const highlightLines = rangeParser(strlineNumbers);
+          const highlight = highlightLines;
+          const data = highlight.includes(applyHighlights) ? 'highlight' : null;
+          return { data };
+        } else {
+          return {};
+        }
+      };
+
+      return hasLang ? (
+        <SyntaxHighlighter
+          style={syntaxTheme}
+          language={hasLang[1]}
+          PreTag="div"
+          className="codeStyle"
+          // showLineNumbers
+          wrapLines={hasMeta}
+          useInlineStyles
+          lineProps={applyHighlights}
+        >
+          {props.children}
+        </SyntaxHighlighter>
+      ) : (
+        <code className={className} {...props} />
+      );
+    },
+  };
+
+  return (
+    <ReactMarkdown
+      components={MarkdownComponents}
+      remarkPlugins={[remarkGfm]}
+      rehypePlugins={[rehypeRaw]}
+    >
+      {children}
+    </ReactMarkdown>
+  );
+};
+
+export default Markdown;
