--- public/sitemap.xml
@@ -1,34 +1,0 @@
-
-    <?xml version="1.0" encoding="UTF-8"?>
-    <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-        
-              <url>
-                  <loc>https://codeedit.app/developer</loc>
-              </url>
-            
-              <url>
-                  <loc>https://codeedit.app/extensions</loc>
-              </url>
-            
-              <url>
-                  <loc>https://codeedit.app</loc>
-              </url>
-            
-              <url>
-                  <loc>https://codeedit.app/resources</loc>
-              </url>
-            
-              <url>
-                  <loc>https://codeedit.app/whats-new</loc>
-              </url>
-            
-              <url>
-                  <loc>https://codeedit.app/blog/test</loc>
-              </url>
-            
-              <url>
-                  <loc>https://codeedit.app/blog/2023/12/redesigned-apple-tv-app-elevates-the-viewing-experience</loc>
-              </url>
-            
-    </urlset>
-
