+++ scripts/generate-sitemap.mjs
@@ -19,6 +19,8 @@ async function generate() {
     '!pages/404.tsx',
   ]);
 
+  console.log(pages);
+
   const sitemap = `
     <?xml version="1.0" encoding="UTF-8"?>
     <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
